Moved lexClient into /channelCapture
